Refactor TabMenu into a component
