+++ src/routes/Package/Package.styled.tsx
@@ -1,7 +1,6 @@
 import styled from "@emotion/styled";
 import metrics from "../../theme/metrics";
 import colours from "../../theme/colours";
-import { TabProps } from "./TabProps";
 
 export const StyledContainer = styled.div({
   width: "50%",
@@ -28,23 +27,3 @@ export const StyledDateCreated = styled.h3({
   color: colours.grey,
   margin: 0,
 });
-
-export const StyledTabMenu = styled.div({
-  marginTop: 25,
-});
-
-export const StyledTabMenuItem = styled.button<TabProps>((props) => ({
-  padding: "5px 10px",
-  border: "none",
-  outline: "none",
-  transition: "ease 0.15s",
-  backgroundColor: props.selected ? colours.primary : "revert",
-  color: props.selected ? colours.white : "revert",
-
-  ":hover": {
-    backgroundColor: colours.primary,
-    transform: "scale(1.05)",
-    color: colours.white,
-    cursor: "pointer",
-  },
-}));
